requirements.txt
Flask
scikit-learn
PyPDF2
python-docx
nltk
keybert
